refactor: segment-related functions
